[2302] [Time Beats: 100.00%] [Memory Beats: 14.11%] - LeetPush

+++ LeetCode/Count-Subarrays-With-Score-Less-Than-K.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     long long countSubarrays(vector<int>& nums, long long k) {
-        long long n = nums.size();
-        long long l = 0;
+        int n = nums.size();
+        int l = 0;
         long long ans = 0;
         long long curSum = 0;
-        for(long long r = 0 ; r < n ; r++){
+        for(int r = 0 ; r < n ; r++){
             curSum += nums[r];
             while(l <= r && curSum*(r-l+1) >= k){
                 curSum -= nums[l];
